Conditional Drop table

+++ db/migrate/20130709020520_drop_virtual_connections_table.rb
@@ -1,6 +1,6 @@
 class DropVirtualConnectionsTable < ActiveRecord::Migration
   def up
-    drop_table :virtual_connections
+    drop_table :virtual_connections if ActiveRecord::Base.connection.table_exists?('virtual_connections')
   end
 
   def down
